# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Logenhelfer API
  version: 1.0.0
  description: Backend-agnostic contract. Rails serves it today, Express later.
paths:
...
  /api/v1/events:
    get:
      operationId: listEvents
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: per_page
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
        - name: sort
          in: query
          schema: { type: string, default: date }
        - name: from
          in: query
          description: Inclusive lower bound on event date (YYYY-MM-DD).
          schema: { type: string, format: date }
        - name: to
          in: query
          description: Inclusive upper bound on event date (YYYY-MM-DD).
          schema: { type: string, format: date }
      responses:
        '200':
          description: Paginated event list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EventList' }
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createEvent
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EventInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Event' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/v1/events/{uuid}:
    parameters:
      - name: uuid
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: getEvent
      responses:
        '200':
          description: Event with participants
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EventWithParticipants' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: updateEvent
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EventInput' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Event' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
    delete:
      operationId: deleteEvent
      responses:
        '204': { description: Soft-deleted }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/external_events:
    get:
      operationId: listExternalEvents
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: per_page
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
        - name: from
          in: query
          description: Inclusive lower bound on event date (YYYY-MM-DD). When omitted, defaults to "yesterday onward".
          schema: { type: string, format: date }
        - name: to
          in: query
          description: Inclusive upper bound on event date (YYYY-MM-DD).
          schema: { type: string, format: date }
      responses:
        '200':
          description: Paginated external event list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExternalEventList' }
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createExternalEvent
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExternalEventInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExternalEvent' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
  /api/v1/external_events/{uuid}:
    parameters:
      - name: uuid
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: getExternalEvent
      responses:
        '200':
          description: External event with participants
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExternalEventWithParticipants' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: updateExternalEvent
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExternalEventInput' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExternalEvent' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
    delete:
      operationId: deleteExternalEvent
      responses:
        '204': { description: Soft-deleted }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
  /api/v1/external_event_ics_sources:
    get:
      operationId: listExternalEventIcsSources
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: per_page
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
        - name: sort
          in: query
          schema: { type: string, default: name }
      responses:
        '200':
          description: Paginated ICS source list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExternalEventIcsSourceList' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: createExternalEventIcsSource
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExternalEventIcsSourceInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExternalEventIcsSource' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/v1/external_event_ics_sources/options:
    get:
      operationId: listExternalEventIcsSourceOptions
      description: >-
        Members-readable {uuid, name} list (no url) for the calendar filter
        dropdown - gated on `index ExternalEvent`, which every authenticated
        user holds, unlike GET /external_event_ics_sources above which is
        gated on `manage ExternalEvent`.
      responses:
        '200':
          description: Unpaginated list of non-deleted ICS source options
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExternalEventIcsSourceOptionList' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/external_event_ics_sources/{uuid}:
    parameters:
      - name: uuid
        in: path
        required: true
        schema: { type: string }
    patch:
      operationId: updateExternalEventIcsSource
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExternalEventIcsSourceUpdate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExternalEventIcsSource' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
    delete:
      operationId: deleteExternalEventIcsSource
      responses:
        '204': { description: Soft-deleted }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/external_event_ics_sources/{uuid}/sync:
    parameters:
      - name: uuid
        in: path
        required: true
        schema: { type: string }
    post:
      operationId: syncExternalEventIcsSourceNow
      responses:
        '200':
          description: Sync result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExternalEventIcsSyncResult' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  responses:
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error: { type: string }
...
    Forbidden:
      description: Authenticated but not permitted
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error: { type: string }
...
    NotFound:
      description: No record with that identifier
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error: { type: string }
    UnprocessableEntity:
      description: Validation failed
      content:
        application/json:
          schema:
            type: object
            required: [error, detail]
            properties:
              error: { type: string }
              detail: { type: string }
  schemas:
...
    Event:
      type: object
      additionalProperties: false
      required: [uuid, title, date, whole_day, location, created_by_id, created_at, updated_at]
      properties:
        uuid: { type: string }
        title: { type: string }
        date: { type: string, format: date }
        time: { type: string, nullable: true, example: '19:00' }
        end_time: { type: string, nullable: true, example: '22:00' }
        whole_day: { type: boolean }
        location: { type: string, nullable: true }
        public_description: { type: string, nullable: true }
        private_description: { type: string, nullable: true }
        created_by_id: { type: integer }
        updated_by_id: { type: integer, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    EventParticipantSummary:
      type: object
      additionalProperties: false
      required: [uuid, fullname]
      properties:
        uuid: { type: string }
        fullname: { type: string }
...
    EventWithParticipants:
      # Flattened rather than `allOf: [Event, {participants}]` - openapi_parser does
      # not enforce a composed schema's top-level `additionalProperties: false`, so
      # an allOf here would silently stop catching a stray top-level field.
      type: object
      additionalProperties: false
      required: [uuid, title, date, whole_day, location, created_by_id, created_at, updated_at, participants]
      properties:
        uuid: { type: string }
        title: { type: string }
        date: { type: string, format: date }
        time: { type: string, nullable: true, example: '19:00' }
        end_time: { type: string, nullable: true, example: '22:00' }
        whole_day: { type: boolean }
        location: { type: string, nullable: true }
        public_description: { type: string, nullable: true }
        private_description: { type: string, nullable: true }
        created_by_id: { type: integer }
        updated_by_id: { type: integer, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        participants:
          type: array
          items: { $ref: '#/components/schemas/EventParticipantSummary' }
    EventList:
      type: object
      additionalProperties: false
      required: [rows, row_count]
      properties:
        rows:
          type: array
          items: { $ref: '#/components/schemas/Event' }
        row_count: { type: integer }
...
    EventInput:
      type: object
      additionalProperties: false
      required: [title, date]
      properties:
        title: { type: string }
        date: { type: string, format: date }
        time: { type: string, nullable: true }
        end_time: { type: string, nullable: true }
        whole_day: { type: boolean }
        location: { type: string, nullable: true }
        public_description: { type: string, nullable: true }
        private_description: { type: string, nullable: true }
...
    ExternalEventParticipant:
      type: object
      additionalProperties: false
      required: [user_uuid, fullname, festive_board, subscription_confirmed]
      properties:
        user_uuid: { type: string }
        fullname: { type: string }
        festive_board: { type: boolean }
        subscription_confirmed: { type: boolean }
    ExternalEvent:
      type: object
      additionalProperties: false
      required: [uuid, title, host, location, date, created_by_id, updated_by_id, created_at, updated_at, ics_source_id, ics_source_uuid]
      properties:
        uuid: { type: string }
        title: { type: string }
        host: { type: string, nullable: true }
        location: { type: string }
        description: { type: string, nullable: true }
        date: { type: string, format: date }
        time: { type: string, nullable: true, example: '19:00' }
        end_time: { type: string, nullable: true, example: '22:00' }
        ics_source_id: { type: integer, nullable: true }
        ics_source_uuid: { type: string, nullable: true, description: 'Resolved from ics_source_id — the public identifier of the ExternalEventIcsSource this event was synced from, or null for a manually-created event.' }
        created_by_id: { type: integer }
        updated_by_id: { type: integer, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ExternalEventWithParticipants:
      type: object
      additionalProperties: false
      required: [uuid, title, host, location, date, created_by_id, updated_by_id, created_at, updated_at, ics_source_id, ics_source_uuid, participants]
      properties:
        uuid: { type: string }
        title: { type: string }
        host: { type: string, nullable: true }
        location: { type: string }
        description: { type: string, nullable: true }
        date: { type: string, format: date }
        time: { type: string, nullable: true, example: '19:00' }
        end_time: { type: string, nullable: true, example: '22:00' }
        ics_source_id: { type: integer, nullable: true }
        ics_source_uuid: { type: string, nullable: true, description: 'Resolved from ics_source_id — the public identifier of the ExternalEventIcsSource this event was synced from, or null for a manually-created event.' }
        created_by_id: { type: integer }
        updated_by_id: { type: integer, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        participants:
          type: array
          items: { $ref: '#/components/schemas/ExternalEventParticipant' }
    ExternalEventList:
      type: object
      additionalProperties: false
      required: [rows, row_count]
      properties:
        rows:
          type: array
          items: { $ref: '#/components/schemas/ExternalEvent' }
        row_count: { type: integer }
    ExternalEventInput:
      type: object
      additionalProperties: false
      properties:
        title: { type: string }
        host: { type: string }
        location: { type: string }
        description: { type: string, nullable: true }
        date: { type: string, format: date }
        time: { type: string }
        end_time: { type: string }
...
    ExternalEventIcsSource:
      type: object
      additionalProperties: false
      required: [uuid, name, url, created_at]
      properties:
        uuid: { type: string }
        name: { type: string }
        url: { type: string }
        created_at: { type: string, format: date-time }
    ExternalEventIcsSourceList:
      type: object
      additionalProperties: false
      required: [rows, row_count]
      properties:
        rows:
          type: array
          items: { $ref: '#/components/schemas/ExternalEventIcsSource' }
        row_count: { type: integer }
    ExternalEventIcsSourceInput:
      type: object
      additionalProperties: false
      required: [name, url]
      properties:
        name: { type: string }
        url: { type: string }
    ExternalEventIcsSourceUpdate:
      type: object
      additionalProperties: false
      properties:
        name: { type: string, maxLength: 255 }
        url: { type: string, maxLength: 255 }
    ExternalEventIcsSourceOption:
      type: object
      additionalProperties: false
      required: [uuid, name]
      properties:
        uuid: { type: string }
        name: { type: string }
    ExternalEventIcsSourceOptionList:
      type: object
      additionalProperties: false
      required: [rows]
      properties:
        rows:
          type: array
          items: { $ref: '#/components/schemas/ExternalEventIcsSourceOption' }
    ExternalEventIcsSyncResult:
      type: object
      additionalProperties: false
      required: [created, updated, removed]
      properties:
        created: { type: integer }
        updated: { type: integer }
        removed: { type: integer }
...
security:
  - bearerAuth: []